Fix #291.
Fix some issues with point labeling using the 'EightSurrounding' algorithm.

* the text labels were not positioned correctly around the symbol (due to a regression)
* the code did not correctly handle the case with Y-down (GD renderer)
* there was a memory leak - one SE_RenderPointStyle per label
* added checks on the geometry type and symbol


git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@2475 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/Stylization/SE_LineBuffer.cpp
@@ -354,7 +354,7 @@ void SE_LineBuffer::PopulateXFBuffer(bool isPolygon)
     }
 
     if (outline->point_count())
-        outline->SetGeometryType(isPolygon? (int)FdoGeometryType_Polygon : (int)FdoGeometryType_LineString);
+        outline->SetGeometryType(isPolygon? FdoGeometryType_Polygon : FdoGeometryType_LineString);
 
     if (m_xf_weight >= 2.0)
     {
